Fixes IPublishedShapshot and friends
