+++ src/Umbraco.Tests/TestHelpers/TestObjects-Mocks.cs
@@ -106,7 +106,7 @@ namespace Umbraco.Tests.TestHelpers
         {
             var httpContext = Mock.Of<HttpContextBase>();
 
-            var publishedSnapshotMock = new Mock<IPublishedShapshot>();
+            var publishedSnapshotMock = new Mock<IPublishedSnapshot>();
             publishedSnapshotMock.Setup(x => x.Members).Returns(Mock.Of<IPublishedMemberCache>());
             var publishedSnapshot = publishedSnapshotMock.Object;
             var publishedSnapshotServiceMock = new Mock<IPublishedSnapshotService>();
